requirements.txt
matplotlib==3.8.3
python-dotenv==1.0.1
